crawler/requirements.txt
boto3==1.38.30
Jinja2==3.1.6
njsparser==2.16
Pillow==11.2.1
psycopg[binary]==3.3.3
python-dotenv==1.2.2
resend==2.26.0
scrapy==2.14.2
scrapy-seleniumbase-cdp==2.0.0
seleniumbase==4.47.8
